+++ Electron/src/app/core/services/index.ts
@@ -1,2 +1,1 @@
 export * from './electron/electron.service';
-export * from './menu/menu.service';