+++ Electron/src/testing/electron-service.stub.ts
@@ -1,5 +1,6 @@
 export const electronServiceStub = {
   isElectron: false,
+  openExternal: () => undefined,
   ipcRenderer: {
     on: () => undefined,
     send: () => undefined